refactor: 코드 리뷰 중 요청된 내용 반영
